+++ W3Edit.W3Speech/W3Edit.W3Speech.csproj
@@ -44,7 +44,8 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="W3SpeechFile.cs" />
+    <Compile Include="Coder.cs" />
+    <Compile Include="DataStructures.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
